requirements.txt
asgiref==3.2.7
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.4
Django==3.0.4
django-reversion==3.0.7
django-storages==1.9.1
djangorestframework==3.11.0
ecdsa==0.15
google-api-core==1.16.0
google-cloud-core==1.3.0
django-filter
django-tables2
django-nucleus
google-cloud-storage
google-resumable-media
idna==2.9
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-jose==3.1.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
uWSGI==2.0.18
django-mathfilters
django-storages[google]
stripe
django-allauth
django-crispy-forms
analytics-python
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-pubsub==1.4.2
django_select2
oauth2client
google_auth_oauthlib
sentry-sdk==0.10.2
django-notifications-hq
sendgrid==5.6.0
